move pydantic util code to util module
